regex: Organizes the classes per file
moves Pcre to backends/pcre.cc
moves PegexMatch to regex_match.h
