+++ src/regex/backend/pcre.cc
@@ -13,23 +13,135 @@
  *
  */
 
-#include <pcre.h>
 
 #include <iostream>
 #include <fstream>
 #include <string>
 #include <list>
+#include <pcre.h>
 
-#ifndef SRC_REGEX_BACKEND_PCRE_H_
-#define SRC_REGEX_BACKEND_PCRE_H_
+#include "src/regex/backend/pcre.h"
+#include "src/regex/regex_match.h"
+
 
 namespace modsecurity {
 namespace regex {
+namespace backend {
 
 
+#if PCRE_HAVE_JIT
+#define pcre_study_opt PCRE_STUDY_JIT_COMPILE
+#else
+#define pcre_study_opt 0
+#endif
 
+
+Pcre::Pcre(const std::string& pattern_)
+    : pattern(pattern_.empty() ? ".*" : pattern_) {
+    const char *errptr = NULL;
+    int erroffset;
+
+    m_pc = pcre_compile(pattern.c_str(), PCRE_DOTALL|PCRE_MULTILINE,
+        &errptr, &erroffset, NULL);
+
+    m_pce = pcre_study(m_pc, pcre_study_opt, &errptr);
+}
+
+
+Pcre::~Pcre() {
+    if (m_pc != NULL) {
+        pcre_free(m_pc);
+        m_pc = NULL;
+    }
+    if (m_pce != NULL) {
+#if PCRE_HAVE_JIT
+        pcre_free_study(m_pce);
+#else
+        pcre_free(m_pce);
+#endif
+        m_pce = NULL;
+    }
+}
+
+
+std::list<RegexMatch> Pcre::searchAll(const std::string& s) const {
+    const char *subject = s.c_str();
+    const std::string tmpString = std::string(s.c_str(), s.size());
+    int ovector[OVECCOUNT];
+    int rc, i, offset = 0;
+    std::list<RegexMatch> retList;
+
+    do {
+        rc = pcre_exec(m_pc, m_pce, subject,
+            s.size(), offset, 0, ovector, OVECCOUNT);
+
+        for (i = 0; i < rc; i++) {
+            size_t start = ovector[2*i];
+            size_t end = ovector[2*i+1];
+            size_t len = end - start;
+            if (end > s.size()) {
+                rc = 0;
+                break;
+            }
+            std::string match = std::string(tmpString, start, len);
+            offset = start + len;
+            retList.push_front(RegexMatch(match, start));
+
+            if (len == 0) {
+                rc = 0;
+                break;
+            }
+        }
+    } while (rc > 0);
+
+    return retList;
+}
+
+
+bool Pcre::searchOneMatch(const std::string& s, std::vector<RegexMatchCapture>& captures) const {
+    const char *subject = s.c_str();
+    int ovector[OVECCOUNT];
+
+    int rc = pcre_exec(m_pc, m_pce, subject, s.size(), 0, 0, ovector, OVECCOUNT);
+
+    for (int i = 0; i < rc; i++) {
+        size_t start = ovector[2*i];
+        size_t end = ovector[2*i+1];
+        size_t len = end - start;
+        if (end > s.size()) {
+            continue;
+        }
+        RegexMatchCapture capture(i, start, len);
+        captures.push_back(capture);
+    }
+
+    return (rc > 0);
+}
+
+
+int Pcre::search(const std::string& s, RegexMatch *match) const {
+    int ovector[OVECCOUNT];
+    int ret = pcre_exec(m_pc, m_pce, s.c_str(),
+        s.size(), 0, 0, ovector, OVECCOUNT) > 0;
+
+    if (ret > 0) {
+        *match = RegexMatch(
+            std::string(s, ovector[ret-1], ovector[ret] - ovector[ret-1]),
+            0);
+    }
+
+    return ret;
+}
+
+
+int Pcre::search(const std::string& s) const {
+    int ovector[OVECCOUNT];
+    return pcre_exec(m_pc, m_pce, s.c_str(),
+        s.size(), 0, 0, ovector, OVECCOUNT) > 0;
+}
+
+
+}  // namespace backend
 }  // namespace regex
 }  // namespace modsecurity
 
-
-#endif  // SRC_REGEX_BACKEND_PCRE_H_
